Create BaseServlet Created a new base servlet class that is used as a superclass for all the servlets that require access to ConversationStore, MessageStore and UserStore.
