+++ src/main/java/codeu/controller/ProfileServlet.java
@@ -3,9 +3,6 @@ package codeu.controller;
 import codeu.model.data.Conversation;
 import codeu.model.data.Message;
 import codeu.model.data.User;
-import codeu.model.store.basic.ConversationStore;
-import codeu.model.store.basic.MessageStore;
-import codeu.model.store.basic.UserStore;
 import java.io.IOException;
 import java.time.Instant;
 import java.util.List;
@@ -18,49 +15,7 @@ import org.jsoup.Jsoup;
 import org.jsoup.safety.Whitelist;
 
 /** Servlet class responsible for the profile page. */
-public class ProfileServlet extends HttpServlet {
-
-  /** Store class that gives access to Conversations. */
-  private ConversationStore conversationStore;
-
-  /** Store class that gives access to Messages. */
-  private MessageStore messageStore;
-
-  /** Store class that gives access to Users. */
-  private UserStore userStore;
-
-  /** Set up state for handling chat requests. */
-  @Override
-  public void init() throws ServletException {
-    super.init();
-    setConversationStore(ConversationStore.getInstance());
-    setMessageStore(MessageStore.getInstance());
-    setUserStore(UserStore.getInstance());
-  }
-
-  /**
-   * Sets the ConversationStore used by this servlet. This function provides a common setup method
-   * for use by the test framework or the servlet's init() function.
-   */
-  void setConversationStore(ConversationStore conversationStore) {
-    this.conversationStore = conversationStore;
-  }
-
-  /**
-   * Sets the MessageStore used by this servlet. This function provides a common setup method for
-   * use by the test framework or the servlet's init() function.
-   */
-  void setMessageStore(MessageStore messageStore) {
-    this.messageStore = messageStore;
-  }
-
-  /**
-   * Sets the UserStore used by this servlet. This function provides a common setup method for use
-   * by the test framework or the servlet's init() function.
-   */
-  void setUserStore(UserStore userStore) {
-    this.userStore = userStore;
-  }
+public class ProfileServlet extends BaseServlet {
 
   /**
    * This function fires when a user navigates to the profile page. It gets the conversation title from
