+++ src/main/java/codeu/controller/TestDataServlet.java
@@ -14,59 +14,13 @@
 
 package codeu.controller;
 
-import codeu.model.store.basic.ConversationStore;
-import codeu.model.store.basic.MessageStore;
-import codeu.model.store.basic.UserStore;
 import java.io.IOException;
 import javax.servlet.ServletException;
-import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 /** Servlet class responsible for loading test data. */
-public class TestDataServlet extends HttpServlet {
-
-  /** Store class that gives access to Conversations. */
-  private ConversationStore conversationStore;
-
-  /** Store class that gives access to Messages. */
-  private MessageStore messageStore;
-
-  /** Store class that gives access to Users. */
-  private UserStore userStore;
-
-  /** Set up state for handling the load test data request. */
-  @Override
-  public void init() throws ServletException {
-    super.init();
-    setConversationStore(ConversationStore.getInstance());
-    setMessageStore(MessageStore.getInstance());
-    setUserStore(UserStore.getInstance());
-  }
-
-  /**
-   * Sets the ConversationStore used by this servlet. This function provides a common setup method
-   * for use by the test framework or the servlet's init() function.
-   */
-  void setConversationStore(ConversationStore conversationStore) {
-    this.conversationStore = conversationStore;
-  }
-
-  /**
-   * Sets the MessageStore used by this servlet. This function provides a common setup method for
-   * use by the test framework or the servlet's init() function.
-   */
-  void setMessageStore(MessageStore messageStore) {
-    this.messageStore = messageStore;
-  }
-
-  /**
-   * Sets the UserStore used by this servlet. This function provides a common setup method for use
-   * by the test framework or the servlet's init() function.
-   */
-  void setUserStore(UserStore userStore) {
-    this.userStore = userStore;
-  }
+public class TestDataServlet extends BaseServlet {
 
   /**
    * This function fires when a user requests the /testdata URL. It simply forwards the request to
